+++ specs/resources/markets.spec.ts
@@ -34,6 +34,8 @@ describe('Markets resource', () => {
 			tax_calculator: cl.avalara_accounts.relationship(TestData.id),
 			customer_group: cl.customer_groups.relationship(TestData.id),
 			geocoder: cl.geocoders.relationship(TestData.id),
+			default_shipping_method: cl.shipping_methods.relationship(TestData.id),
+			default_payment_method: cl.payment_methods.relationship(TestData.id),
 		}
 
     const attributes = { ...createAttributes, reference: TestData.reference }
@@ -374,6 +376,48 @@ describe('Markets resource', () => {
 	
 	})
 	/* relationship.geocoder stop */
+	
+
+	/* relationship.default_shipping_method start */
+	it(resourceType + '.default_shipping_method', async () => {
+	
+		const id = TestData.id
+		const params = { fields: { shipping_methods: CommonData.paramsFields } }
+	
+		const intId = cl.addRequestInterceptor((request) => {
+			expect(request.options.method).toBe('GET')
+			checkCommon(request, resourcePath, id, currentAccessToken, 'default_shipping_method')
+			checkCommonParams(request, params)
+			return interceptRequest()
+		})
+	
+		await cl[resourcePath].default_shipping_method(id, params, CommonData.options)
+			.catch(handleError)
+			.finally(() => cl.removeInterceptor('request'))
+	
+	})
+	/* relationship.default_shipping_method stop */
+	
+
+	/* relationship.default_payment_method start */
+	it(resourceType + '.default_payment_method', async () => {
+	
+		const id = TestData.id
+		const params = { fields: { payment_methods: CommonData.paramsFields } }
+	
+		const intId = cl.addRequestInterceptor((request) => {
+			expect(request.options.method).toBe('GET')
+			checkCommon(request, resourcePath, id, currentAccessToken, 'default_payment_method')
+			checkCommonParams(request, params)
+			return interceptRequest()
+		})
+	
+		await cl[resourcePath].default_payment_method(id, params, CommonData.options)
+			.catch(handleError)
+			.finally(() => cl.removeInterceptor('request'))
+	
+	})
+	/* relationship.default_payment_method stop */
 	
 
 	/* relationship.stores start */