+++ specs/resources/shipments.spec.ts
@@ -374,6 +374,27 @@ describe('Shipments resource', () => {
 	/* relationship.delivery_lead_time stop */
 	
 
+	/* relationship.pickup start */
+	it(resourceType + '.pickup', async () => {
+	
+		const id = TestData.id
+		const params = { fields: { pickups: CommonData.paramsFields } }
+	
+		const intId = cl.addRequestInterceptor((request) => {
+			expect(request.options.method).toBe('GET')
+			checkCommon(request, resourcePath, id, currentAccessToken, 'pickup')
+			checkCommonParams(request, params)
+			return interceptRequest()
+		})
+	
+		await cl[resourcePath].pickup(id, params, CommonData.options)
+			.catch(handleError)
+			.finally(() => cl.removeInterceptor('request'))
+	
+	})
+	/* relationship.pickup stop */
+	
+
 	/* relationship.stock_line_items start */
 	it(resourceType + '.stock_line_items', async () => {
 	
@@ -921,5 +942,57 @@ describe('Shipments resource', () => {
 	
 	})
 	/* trigger._purchase stop */
+	
+
+	/* trigger._add_tags start */
+	it(resourceType + '._add_tags', async () => {
+	
+		let triggerAttr = '_add_tags'
+		if (!triggerAttr.startsWith('_')) triggerAttr = `_${triggerAttr}`
+	
+		const triggerValue = randomValue('string')
+		const attributes = { [triggerAttr]: triggerValue }
+	  const id = TestData.id
+	
+		const intId = cl.addRequestInterceptor((request) => {
+			const data = JSON.parse(String(request.options.body))
+			expect(request.options.method).toBe('PATCH')
+			checkCommon(request, resourcePath, id, currentAccessToken)
+			checkCommonData(data, resourceType, attributes, id)
+			return interceptRequest()
+		})
+	
+		await cl[resourcePath]._add_tags(id, triggerValue, {}, CommonData.options)
+			.catch(handleError)
+			.finally(() => cl.removeInterceptor('request'))
+	
+	})
+	/* trigger._add_tags stop */
+	
+
+	/* trigger._remove_tags start */
+	it(resourceType + '._remove_tags', async () => {
+	
+		let triggerAttr = '_remove_tags'
+		if (!triggerAttr.startsWith('_')) triggerAttr = `_${triggerAttr}`
+	
+		const triggerValue = randomValue('string')
+		const attributes = { [triggerAttr]: triggerValue }
+	  const id = TestData.id
+	
+		const intId = cl.addRequestInterceptor((request) => {
+			const data = JSON.parse(String(request.options.body))
+			expect(request.options.method).toBe('PATCH')
+			checkCommon(request, resourcePath, id, currentAccessToken)
+			checkCommonData(data, resourceType, attributes, id)
+			return interceptRequest()
+		})
+	
+		await cl[resourcePath]._remove_tags(id, triggerValue, {}, CommonData.options)
+			.catch(handleError)
+			.finally(() => cl.removeInterceptor('request'))
+	
+	})
+	/* trigger._remove_tags stop */
 	
 })